Add configure.sh file to compile Cython file inside docker image
`configure.sh` had been removed as it was expected to not be required
anymore. However, since `deploy.sh` mounts the local folder inside the
docker image, some files that are created inside the docker image
are hidden under that mounted folder, and cannot be found. To work
around this issue, one need to recreate these files inside the local
mounted folder.